feat(ContextMenu): add NetworkSwitcherContextMenu component and integrate it
