+++ src/Components/Reusable/index.ts
@@ -83,3 +83,4 @@ export { default as AccountFiatBalance } from "./AccountFiatBalance"
 export * from "./GasFeeSpeed"
 export * from "./Banners"
 export * from "./SectionListSeparator"
+export * from "./ContextMenu"
