+++ src/Screens/Flows/App/HomeScreen/Components/Header/Header.tsx
@@ -1,25 +1,16 @@
 import { useNavigation } from "@react-navigation/native"
-import React, { memo, useCallback, useMemo } from "react"
+import { NativeStackNavigationProp } from "@react-navigation/native-stack"
+import React, { memo, useCallback } from "react"
+import { VeWorldLogoSVG } from "~Assets"
+import { BaseIcon, BaseSpacer, BaseText, BaseView, HeaderStyleV2, useWalletConnect } from "~Components"
+import { NetworkSwitcherContextMenu } from "~Components/Reusable/ContextMenu"
+import { SelectedNetworkViewer } from "~Components/Reusable/SelectedNetworkViewer"
+import { ERROR_EVENTS, ScanTarget } from "~Constants"
 import { useBlockchainNetwork, useCameraBottomSheet, useCopyClipboard, useTheme, useVisitedUrls } from "~Hooks"
-import { BaseIcon, BaseSpacer, BaseText, BaseView, useWalletConnect, HeaderStyleV2 } from "~Components"
 import { useI18nContext } from "~i18n"
 import { RootStackParamListHome, Routes, TabStackParamList } from "~Navigation"
 import HapticsService from "~Services/HapticsService"
-import { ERROR_EVENTS, ScanTarget } from "~Constants"
-import { SelectedNetworkViewer } from "~Components/Reusable/SelectedNetworkViewer"
 import { AddressUtils, debug, URIUtils, WalletConnectUtils } from "~Utils"
-import { NativeStackNavigationProp } from "@react-navigation/native-stack"
-import { VeWorldLogoSVG } from "~Assets"
-import {
-    changeSelectedNetwork,
-    clearNFTCache,
-    selectNetworks,
-    selectSelectedNetwork,
-    useAppDispatch,
-    useAppSelector,
-} from "~Storage/Redux"
-import ContextMenu, { ContextMenuAction, ContextMenuOnPressNativeEvent } from "react-native-context-menu-view"
-import { NativeSyntheticEvent } from "react-native"
 
 type Navigation = NativeStackNavigationProp<TabStackParamList, "HomeStack"> &
     NativeStackNavigationProp<RootStackParamListHome, Routes.HOME>
@@ -31,9 +22,6 @@ export const Header = memo(() => {
     const { addVisitedUrl } = useVisitedUrls()
     const { isMainnet } = useBlockchainNetwork()
     const { onPair } = useWalletConnect()
-    const dispatch = useAppDispatch()
-    const networks = useAppSelector(selectNetworks)
-    const currentNetwork = useAppSelector(selectSelectedNetwork)
 
     const { onCopyToClipboard } = useCopyClipboard()
 
@@ -85,35 +73,6 @@ export const Header = memo(() => {
         nav.navigate(Routes.SETTINGS_NETWORK)
     }, [nav])
 
-    const renderNetworkOptions: ContextMenuAction[] = useMemo(() => {
-        const networkOptions = networks.map(net => ({
-            title: net.name,
-            subtitle: net.currentUrl,
-            selected: net.id === currentNetwork.id,
-        }))
-        return [
-            ...networkOptions,
-            {
-                title: LL.NETWORK_ADD_CUSTOM_NODE(),
-                selected: false,
-            },
-        ] as ContextMenuAction[]
-    }, [LL, networks, currentNetwork.id])
-
-    const onContextMenuPress = useCallback(
-        (e: NativeSyntheticEvent<ContextMenuOnPressNativeEvent>) => {
-            HapticsService.triggerImpact({ level: "Light" })
-            if (e.nativeEvent.index === networks.length) {
-                goToChooseNetwork()
-                return
-            }
-
-            dispatch(clearNFTCache())
-            dispatch(changeSelectedNetwork(networks[e.nativeEvent.index]))
-        },
-        [dispatch, networks, goToChooseNetwork],
-    )
-
     return (
         <BaseView w={100} style={HeaderStyleV2}>
             <BaseView flexDirection="row" alignItems="center" alignSelf="center">
@@ -144,10 +103,7 @@ export const Header = memo(() => {
                     testID="HomeScreen_WalletManagementButton"
                 />
                 <BaseSpacer width={8} />
-                <ContextMenu
-                    previewBackgroundColor="transparent"
-                    actions={renderNetworkOptions}
-                    onPress={onContextMenuPress}>
+                <NetworkSwitcherContextMenu>
                     <SelectedNetworkViewer />
                     {isMainnet && (
                         <BaseIcon
@@ -159,7 +115,7 @@ export const Header = memo(() => {
                             haptics="Light"
                         />
                     )}
-                </ContextMenu>
+                </NetworkSwitcherContextMenu>
             </BaseView>
 
             {RenderCameraModal}
