+++ src/Screens/Flows/App/NFT/NFTScreen/Components/NftScreenHeader.tsx
@@ -7,6 +7,7 @@ import {
     SelectedNetworkViewer,
     HeaderStyleV2,
 } from "~Components"
+import { NetworkSwitcherContextMenu } from "~Components/Reusable/ContextMenu"
 import { useI18nContext } from "~i18n"
 
 type Props = {
@@ -20,7 +21,9 @@ export const NftScreenHeader = ({ openSelectAccountBottomSheet }: Props) => {
         <BaseView style={HeaderStyleV2}>
             <HeaderTitle testID="nfts_title" title={LL.TITLE_NFTS()} leftIconName="icon-image" />
             <BaseView flexDirection="row" justifyContent="space-between" alignItems="center">
-                <SelectedNetworkViewer />
+                <NetworkSwitcherContextMenu>
+                    <SelectedNetworkViewer />
+                </NetworkSwitcherContextMenu>
                 <BaseSpacer width={8} />
                 <ChangeAccountButtonPill action={openSelectAccountBottomSheet} />
             </BaseView>
